refactor: Group Error files under 📁`Error`
