+++ src/Eleventy.js
@@ -5,8 +5,8 @@ import debugUtil from "debug";
 import TemplateData from "./TemplateData.js";
 import TemplateWriter from "./TemplateWriter.js";
 import EleventyExtensionMap from "./EleventyExtensionMap.js";
-import EleventyErrorHandler from "./EleventyErrorHandler.js";
-import EleventyBaseError from "./EleventyBaseError.js";
+import EleventyErrorHandler from "./Errors/EleventyErrorHandler.js";
+import EleventyBaseError from "./Errors/EleventyBaseError.js";
 import EleventyServe from "./EleventyServe.js";
 import EleventyWatch from "./EleventyWatch.js";
 import EleventyWatchTargets from "./EleventyWatchTargets.js";
