+++ src/Engines/TemplateEngine.js
@@ -1,5 +1,5 @@
 import EleventyExtensionMap from "../EleventyExtensionMap.js";
-import EleventyBaseError from "../EleventyBaseError.js";
+import EleventyBaseError from "../Errors/EleventyBaseError.js";
 
 class TemplateEngineConfigError extends EleventyBaseError {}
