+++ src/Errors/EleventyBaseError.js
