+++ src/Errors/UsingCircularTemplateContentReferenceError.js
@@ -1,4 +1,4 @@
-import EleventyBaseError from "../EleventyBaseError.js";
+import EleventyBaseError from "./EleventyBaseError.js";
 
 class UsingCircularTemplateContentReferenceError extends EleventyBaseError {}
