+++ src/TemplateDataInitialGlobalData.js
@@ -1,7 +1,7 @@
 import semver from "semver";
 import lodash from "@11ty/lodash-custom";
 
-import EleventyBaseError from "./EleventyBaseError.js";
+import EleventyBaseError from "./Errors/EleventyBaseError.js";
 import { getEleventyPackageJson } from "./Util/ImportJsonSync.js";
 
 const { set: lodashSet } = lodash;
