+++ src/TemplateEngineManager.js
@@ -1,4 +1,4 @@
-import EleventyBaseError from "./EleventyBaseError.js";
+import EleventyBaseError from "./Errors/EleventyBaseError.js";
 import { EleventyImportFromEleventy } from "./Util/Require.js";
 
 class TemplateEngineManagerConfigError extends EleventyBaseError {}
