+++ src/TemplateRender.js
@@ -1,6 +1,6 @@
 import { TemplatePath } from "@11ty/eleventy-utils";
 
-import EleventyBaseError from "./EleventyBaseError.js";
+import EleventyBaseError from "./Errors/EleventyBaseError.js";
 import EleventyExtensionMap from "./EleventyExtensionMap.js";
 import CustomEngine from "./Engines/Custom.js";
