+++ src/TemplateWriter.js
@@ -5,9 +5,9 @@ import Template from "./Template.js";
 import TemplateMap from "./TemplateMap.js";
 import EleventyFiles from "./EleventyFiles.js";
 import EleventyExtensionMap from "./EleventyExtensionMap.js";
-import EleventyBaseError from "./EleventyBaseError.js";
-import EleventyErrorHandler from "./EleventyErrorHandler.js";
-import EleventyErrorUtil from "./EleventyErrorUtil.js";
+import EleventyBaseError from "./Errors/EleventyBaseError.js";
+import EleventyErrorHandler from "./Errors/EleventyErrorHandler.js";
+import EleventyErrorUtil from "./Errors/EleventyErrorUtil.js";
 import FileSystemSearch from "./FileSystemSearch.js";
 import ConsoleLogger from "./Util/ConsoleLogger.js";
